jclklib: Remove less -R from Makefile
This commit removes the less -R command from the default target in the
Makefile of client, common and proxy.

The less -R command was originally used to paginate the output of the make
command. However, this caused an issue where the build process would pause
until the user manually quit less by pressing "q".

This is not a typical or recommended use of less, especially in a Makefile.
less is designed for interactive use in a shell to view the contents of a
file or the output of a command page by page. In a Makefile, it can cause
problems if the Makefile is run in a non-interactive environment where
there's no user to scroll through the pages.

Signed-off-by: Song Yoong Siang <yoong.siang.song@intel.com>