jclklib: Rename jclklib.so to libjclk.so in Makefile
This commit updates the Makefile to change the target name from jclklib.so
to libjclk.so. This change aligns with the convention of prefixing shared
library names with 'lib'.

This change ensures that the generated shared library follows the standard
naming convention, which can help prevent confusion and potential issues
with library linking in the future.

Signed-off-by: Song Yoong Siang <yoong.siang.song@intel.com>