jclklib: Remove unnecessary std::move from promise.get_future()
This commit removes an unnecessary use of std::move from
promise.get_future() in the registerWork function of the Transport class.
The std::move was causing a compiler warning because it was applied to a
temporary object, which can prevent certain compiler optimizations.

Removing std::move in this case does not affect the functionality of the
code, as promise.get_future() already returns a temporary object that can
be efficiently moved.

Signed-off-by: Song Yoong Siang <yoong.siang.song@intel.com>