jclklib: Update Makefile to include ptpmgmt library path in LIBS_FLAGS
This commit updates the Makefile for building the jclklib_proxy. It
modifies the LIBS_FLAGS variable to include the path to the ptpmgmt
library using the -L option. This change ensures that the linker can find
the ptpmgmt library when building jclklib_proxy, even if the library is not
installed in the standard library paths.

Signed-off-by: Song Yoong Siang <yoong.siang.song@intel.com>