jclklib: Add Proxy message queue liveness check
Check whether Proxy still alive by sending connect msg and wait for
response.

A null check is added for the dynamic cast of `connectMsg.get()`
to `ClientConnectMessage` in the `check_proxy_liveness` function.

Without null check, if the dynamic cast failed, the program would attempt to
dereference a null pointer, leading to undefined behavior. With this null
check, if the dynamic cast fails, the function will print a debug message
and return false, preventing the null pointer dereference.

Default timeout for proxy liveness is 1s instead of taking
the value of timeout from wait API.

Signed-off-by: Song Yoong Siang <yoong.siang.song@intel.com>