Add null check for dynamic cast in check_proxy_liveness
This commit adds a null check for the dynamic cast of `connectMsg.get()`
to `ClientConnectMessage` in the `check_proxy_liveness` function.

Previously, if the dynamic cast failed, the program would attempt to
dereference a null pointer, leading to undefined behavior. With this change,
if the dynamic cast fails, the function will print a debug message and
return false, preventing the null pointer dereference.

Signed-off-by: Song Yoong Siang <yoong.siang.song@intel.com>