Implement C Wrapper for jclklib Client API
This commit introduces a C wrapper for the jclklib client API. The wrapper
provides functions for creating and destroying a client, connecting and
disconnecting, subscribing to events, and waiting for status updates.

It also defines several structures for managing subscriptions, state, and
event counts. This will allow C applications to interact with the jclklib
library.

Signed-off-by: Song Yoong Siang <yoong.siang.song@intel.com>