clkmgr: Fix Listener's getQueueName() method
Previously, `Listener::getQueueName()` returned the string representation
of the message queue file descriptor, which was incorrect.

This patch introduces a `getName()` getter in the Queue class to properly
retrieve the message queue name, and updates `Listener::getQueueName()`
to use this method. This ensures that `getQueueName()` returns the actual
message queue name as intended.

Signed-off-by: Goh, Wei Sheng <wei.sheng.goh@intel.com>