clkmgr: Enable Manpage generation from Markdown via pandoc.
This commit adds support for generating manpages from Markdown files
using pandoc. The `clkmgr_proxy_cfg.md` file is converted to a manpage
format and placed in the `man` directory. The Makefile is updated to
include this new functionality, and the necessary checks for pandoc
are added to the `configure.ac` file. The manpage generation is
triggered by the `install_clkmgr` target.

Signed-off-by: Goh, Wei Sheng <wei.sheng.goh@intel.com>